AifoilGmsh.jl

Documentation of AifoilGmsh.jl for speeding up the creation of airfoil in GMSH

AirfoilGmsh.create_geofile — Function

create_geofile(filename::String; Reynolds = -1, h0 = -1, leading_edge_points = Int64[], trailing_edge_points = Int64[], chord=1.0, dimension=2, elements = :QUAD)

It is the main function of the package. From a csv file containing the airfoil points it creates a .geo file. The .geo file can be created using the function from_url_to_csv. The user can specify just the file name. ´´´julia create_geofile("naca0012.csv") ´´´ It is also possibile to provide extra arguments such as the Reynolds number and/or the first layer height for a better extimation of the boundary-cell properties. It is possible to overwrite the extimation of the trailing edge and leading edge made by the code providing the relative points numbers. The mesh can be created in 2D or 3D. In 3D case by default are created periodic boundary conditions in the z direction. It is possible to create a mesh with the following options:

|Type of element | Dimension | Symbol | |Quadrilateral | 2D | :QUAD| |Hexaedral | 3D | :HEX| |Triangular | 2D | :TRI| |Thetraedreal | 3D | :TETRA|

AirfoilGmsh.from_url_to_csv — Function

from_url_to_csv(url::String)

Provide an url from arifoiltools.com and it writes a formatted CSV.